Restructured folder for dataset details components
